Refactor: Move TreeSize types to primitives module
Extract TreeSize trait and implementations from commitment.rs to a new
primitives module, following the module.rs + module/ pattern:

- Create types/primitives.rs as index module
- Create types/primitives/tree_size.rs with moved TreeSize types
- Update commitment.rs to focus only on commitment tree data structures
- Update re-exports in main types.rs

This establishes primitives/ as the home for foundational types that
other modules depend on, providing better separation of concerns and
preparing for extraction of other primitive types (BlockHash, Height, etc.).

Directory structure:
types/
├── primitives.rs          # Re-exports from primitives/
├── primitives/
│   └── tree_size.rs       # TreeSize trait + pool-specific implementations
└── commitment.rs          # Commitment tree data structures
